fix: made janus id uint as u32, janus doesn't support u64 - thanks e2es

+++ jarust_plugins/src/common.rs
@@ -12,5 +12,5 @@ pub enum JanusId {
     /// String Identifier
     String(String),
     /// Unsigned Integer Identifier
-    Uint(u64),
+    Uint(u32),
 }
